delete Cards funktioniert

+++ demo/src/main/java/com/example/demo/cards/CardInstanceRepository.java
@@ -1,7 +1,8 @@
 package com.example.demo.cards;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardInstanceRepository extends CrudRepository<CardInstance, Integer> {
+public interface CardInstanceRepository extends CrudRepository<CardInstance, Long> {
 }
